Make pipe consumer fault-tolerant.

+++ crates/iori/src/consumer/pipe.rs
@@ -17,7 +17,7 @@ pub struct PipeConsumer {
     recycle: bool,
 
     next: Arc<AtomicU64>,
-    segments: Arc<Mutex<HashMap<u64, PathBuf>>>,
+    segments: Arc<Mutex<HashMap<u64, Option<PathBuf>>>>,
 }
 
 impl PipeConsumer {
@@ -52,21 +52,29 @@ impl PipeConsumer {
         let next = self.next.clone();
         let segments = self.segments.clone();
         Ok(Some(ConsumerOutput::new(Box::pin(file)).on_finish(
-            move || {
+            move |failed: bool| {
                 Box::pin(async move {
                     // Hold the lock so that no one would be able to write new segments and modify `next`
                     let mut segments = segments.lock().await;
 
-                    // write file path to HashMap
-                    segments.insert(sequence, path);
+                    if failed {
+                        // try to remove segment, but ignore the result
+                        let _ = tokio::fs::remove_file(&path).await;
+                        segments.insert(sequence, None);
+                    } else {
+                        // write file path to HashMap
+                        segments.insert(sequence, Some(path));
+                    }
 
                     if sequence == next.load(Ordering::Relaxed) {
                         while let Some(path) = segments.remove(&next.load(Ordering::Relaxed)) {
-                            // open file and write binary content to stdout
-                            let mut file = std::fs::File::open(&path)?;
-                            let _ = std::io::copy(&mut file, &mut std::io::stdout());
-                            if recycle {
-                                tokio::fs::remove_file(&path).await?;
+                            if let Some(path) = path {
+                                // open file and write binary content to stdout
+                                let mut file = std::fs::File::open(&path)?;
+                                let _ = std::io::copy(&mut file, &mut std::io::stdout());
+                                if recycle {
+                                    tokio::fs::remove_file(&path).await?;
+                                }
                             }
 
                             next.fetch_add(1, Ordering::Relaxed);
